Simplify Entity.AttachedBlips
Revert RelationShipGroup.ToString
Rename Order to Priority in Blip

+++ source/scripting/RelationshipGroup.cs
@@ -85,7 +85,7 @@ namespace GTA
 		}
 		public override string ToString()
 		{
-			return "0x" + ((uint)Hash).ToString("X");
+			return "0x" + Hash.ToString("X");
 		}
 
 		public static implicit operator RelationshipGroup(int source)
